api/requirements.txt
uvicorn
fastapi
python-socketio[asgi]
openai-agents
supabase
tiktoken
pdfminer.six